+++ src/components/layouts/Header/DesktopLayout/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { Path } from '@/app/constants/path';
 import { Container } from '@/components/layouts/Container';
-import { InternalLink } from '@/components/parts/InternalLink';
-import { Label } from '@/components/parts/Label';
+import { InternalLink } from '@/components/parts/Text/InternalLink';
+import { Label } from '@/components/parts/Text/Label';
 
 import { styles } from './styles';
