+++ src/components/parts/Image/Icon/index.tsx
