+++ src/components/parts/Image/QRcode/index.tsx
