+++ src/components/parts/Image/QRcode/styles.ts
