--- src/components/parts/TertiaryButton/index.tsx
@@ -1,44 +1,0 @@
-import React, { useMemo } from 'react';
-
-import { IconPattern } from '@/app/assets/icons';
-import { Color } from '@/app/theme/colors';
-import { Icon } from '@/components/parts/Icon';
-import { Label } from '@/components/parts/Label';
-
-import { styles } from './styles';
-
-type ButtonProps = Pick<React.ComponentProps<'button'>, 'onClick' | 'disabled'>;
-
-type Props = {
-  children: string;
-  iconPattern?: IconPattern;
-} & ButtonProps;
-
-export const TertiaryButton: React.FC<Props> = ({
-  children,
-  disabled,
-  iconPattern,
-  onClick,
-}) => {
-  const contentColor = useMemo((): Color => {
-    return disabled ? 'tapa' : 'mirage';
-  }, [disabled]);
-
-  return (
-    <button
-      type="button"
-      disabled={disabled}
-      onClick={onClick}
-      css={styles.container}
-    >
-      <div css={styles.innerWrapper}>
-        {!!iconPattern && (
-          <Icon pattern={iconPattern} stroke={contentColor} size="s" />
-        )}
-        <Label size="s" weight="bold" color={contentColor}>
-          {children}
-        </Label>
-      </div>
-    </button>
-  );
-};
