+++ src/components/parts/Text/InternalLink/styles.ts
