+++ src/stories/components/parts/Button/PrimaryButton/index.stories.tsx
@@ -1,0 +1,16 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { PrimaryButton } from '@/components/parts/Button/PrimaryButton';
+
+export default {
+  component: PrimaryButton,
+} as ComponentMeta<typeof PrimaryButton>;
+
+export const Index: ComponentStoryObj<typeof PrimaryButton> = {
+  args: {
+    children: 'Example',
+    size: 'm',
+    iconPattern: undefined,
+    disabled: false,
+  },
+};
