+++ src/stories/components/parts/Button/SecondaryButton/index.stories.tsx
@@ -1,0 +1,16 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { SecondaryButton } from '@/components/parts/Button/SecondaryButton';
+
+export default {
+  component: SecondaryButton,
+} as ComponentMeta<typeof SecondaryButton>;
+
+export const Index: ComponentStoryObj<typeof SecondaryButton> = {
+  args: {
+    children: 'Example',
+    size: 'm',
+    iconPattern: undefined,
+    disabled: false,
+  },
+};
