+++ src/stories/components/parts/Button/TertiaryButton/index.stories.tsx
@@ -1,0 +1,15 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { TertiaryButton } from '@/components/parts/Button/TertiaryButton';
+
+export default {
+  component: TertiaryButton,
+} as ComponentMeta<typeof TertiaryButton>;
+
+export const Index: ComponentStoryObj<typeof TertiaryButton> = {
+  args: {
+    children: 'Example',
+    iconPattern: undefined,
+    disabled: false,
+  },
+};
