--- src/stories/components/parts/Icon/index.stories.tsx
@@ -1,16 +1,0 @@
-import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-
-import { Icon } from '@/components/parts/Icon';
-
-export default {
-  component: Icon,
-} as ComponentMeta<typeof Icon>;
-
-export const Index: ComponentStoryObj<typeof Icon> = {
-  args: {
-    pattern: 'X',
-    size: 'm',
-    stroke: 'mirage',
-    fill: 'transparent',
-  },
-};
