+++ src/stories/components/parts/Image/Icon/index.stories.tsx
@@ -1,0 +1,16 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { Icon } from '@/components/parts/Image/Icon';
+
+export default {
+  component: Icon,
+} as ComponentMeta<typeof Icon>;
+
+export const Index: ComponentStoryObj<typeof Icon> = {
+  args: {
+    pattern: 'X',
+    size: 'm',
+    stroke: 'mirage',
+    fill: 'transparent',
+  },
+};
