+++ src/stories/components/parts/Image/QRcode/index.stories.tsx
@@ -1,0 +1,21 @@
+import { css } from '@emotion/react';
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+import React from 'react';
+
+import { QRcode } from '@/components/parts/Image/QRcode';
+import { DUMMY_URL } from '@/stories/dummy/url';
+
+export default {
+  component: QRcode,
+} as ComponentMeta<typeof QRcode>;
+
+export const Index: ComponentStoryObj<typeof QRcode> = {
+  args: {
+    url: DUMMY_URL,
+  },
+  render: (args) => (
+    <div css={css({ width: 300 })}>
+      <QRcode {...args} />
+    </div>
+  ),
+};
