+++ src/stories/components/parts/Input/InputField/index.stories.tsx
@@ -1,0 +1,14 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { InputField } from '@/components/parts/Input/InputField';
+
+export default {
+  component: InputField,
+} as ComponentMeta<typeof InputField>;
+
+export const Index: ComponentStoryObj<typeof InputField> = {
+  args: {
+    placeholder: 'Search',
+    onSubmit: (e) => console.log(e.currentTarget.value),
+  },
+};
