--- src/stories/components/parts/InputField/index.stories.tsx
@@ -1,14 +1,0 @@
-import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-
-import { InputField } from '@/components/parts/InputField';
-
-export default {
-  component: InputField,
-} as ComponentMeta<typeof InputField>;
-
-export const Index: ComponentStoryObj<typeof InputField> = {
-  args: {
-    placeholder: 'Search',
-    onSubmit: (e) => console.log(e.currentTarget.value),
-  },
-};
