--- src/stories/components/parts/PrimaryButton/index.stories.tsx
@@ -1,16 +1,0 @@
-import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-
-import { PrimaryButton } from '@/components/parts/PrimaryButton';
-
-export default {
-  component: PrimaryButton,
-} as ComponentMeta<typeof PrimaryButton>;
-
-export const Index: ComponentStoryObj<typeof PrimaryButton> = {
-  args: {
-    children: 'Example',
-    size: 'm',
-    iconPattern: undefined,
-    disabled: false,
-  },
-};
