--- src/stories/components/parts/TertiaryButton/index.stories.tsx
@@ -1,15 +1,0 @@
-import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-
-import { TertiaryButton } from '@/components/parts/TertiaryButton';
-
-export default {
-  component: TertiaryButton,
-} as ComponentMeta<typeof TertiaryButton>;
-
-export const Index: ComponentStoryObj<typeof TertiaryButton> = {
-  args: {
-    children: 'Example',
-    iconPattern: undefined,
-    disabled: false,
-  },
-};
