+++ src/stories/components/parts/Text/Body/index.stories.tsx
@@ -1,0 +1,28 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import { Body } from '@/components/parts/Text/Body';
+
+export default {
+  component: Body,
+} as ComponentMeta<typeof Body>;
+
+const Index: ComponentStoryObj<typeof Body> = {
+  args: {
+    children: 'Example',
+    color: 'mirage',
+  },
+};
+
+export const Fixed = {
+  args: {
+    ...Index.args,
+    size: 's',
+  },
+};
+
+export const Fluid = {
+  args: {
+    ...Index.args,
+    mqSizes: ['s', 'xl'],
+  },
+};
