+++ src/stories/components/patterns/Modal/BasicModal/index.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
 import React, { useState } from 'react';
 
-import { SecondaryButton } from '@/components/parts/SecondaryButton';
+import { SecondaryButton } from '@/components/parts/Button/SecondaryButton';
 import { BasicModal } from '@/components/patterns/Modal/BasicModal';
 
 export default {
